e2e_test/source_inline/requirements.txt
nats-py==2.9.0
psycopg[binary]
psycopg2-binary==2.9.10
pulsar-client==3.6.1
